+++ db/migrate/20240318174956_add_adopter_foster_account_id_to_adopter_applications.rb
@@ -1,6 +1,6 @@
 class AddAdopterFosterAccountIdToAdopterApplications < ActiveRecord::Migration[7.0]
   def change
-    add_column :adopter_applications, :adopter_foster_account_id, :integer
+    add_column :adopter_applications, :adopter_foster_account_id, :bigint
     add_index :adopter_applications, :adopter_foster_account_id
   end
 end
