+++ db/migrate/20240318175438_add_adopter_foster_account_id_to_matches.rb
@@ -1,6 +1,6 @@
 class AddAdopterFosterAccountIdToMatches < ActiveRecord::Migration[7.0]
   def change
-    add_column :matches, :adopter_foster_account_id, :integer
+    add_column :matches, :adopter_foster_account_id, :bigint
     add_index :matches, :adopter_foster_account_id
   end
 end
